# docs/swagger.yaml
definitions:
  github_com_Orange-Health_citadel_apps_task_metadata_structures.TaskMetadata:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      doctor_name:
        type: string
      doctor_notes:
        type: string
      doctor_number:
        type: string
      id:
        type: integer
      is_critical:
        type: boolean
      is_morphle:
        type: boolean
      is_package:
        type: boolean
      task_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  models.TaskMetadata:
    type: object
...
  structures.AuditLog:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      id:
        type: integer
      log:
        allOf:
        - $ref: '#/definitions/structures.Log'
        description: |-
          The log of the AuditLog.
          example: {"log": {"request_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "order_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "sample_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "test_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "investigation_logs": [{"investigation_name": "CBC", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]}]}]}]}]}}
      log_type:
        description: |-
          The log type of the AuditLog.
          example: "Task"
        type: string
      task_id:
        description: |-
          The task ID of the AudiLog.
          example: 1
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
  structures.CommonAPIResponse:
    properties:
      data:
        description: The data of the response.
      error:
        description: |-
          The error of the response.
          example: "Error while fetching"
        type: string
      message:
        description: |-
          The message of the response.
          example: "Fetched Successfully"
        type: string
      validation_errors:
        additionalProperties: true
        description: |-
          Validation Errors
          example: {"field": "error message"}
        type: object
    type: object
  structures.CommonError:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  structures.DeltaValuesStructResponse:
    properties:
      approved_at:
        type: string
      investigation_value:
        type: string
      master_investigation_id:
        type: integer
    type: object
  structures.InvestigationData:
    properties:
      data:
        description: |-
          Data contains the investigation data in text format
          swagger:strfmt binary
        type: string
      data_type:
        description: |-
          The investigation data type.
          example: "medical_remark"
        type: string
      id:
        description: |-
          The investigation data ID.
          example: 1
        type: integer
      investigation_result_id:
        description: |-
          InvestigationResultId is the ID of the investigation result.
          example: 1
        type: integer
    type: object
  structures.InvestigationLogs:
    properties:
      investigation_name:
        description: |-
          The investigation name of the InvestigationLogs.
          example: "CBC"
        type: string
      status_logs:
        description: |-
          The status logs of the InvestigationLogs.
          example: [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.StatusLog'
        type: array
    type: object
  structures.InvestigationResult:
    properties:
      abnormality:
        description: |-
          The investigation abnormality.
          example: "N"
        type: string
      approved_at:
        description: |-
          The investigation Approved At.
          example: "2021-07-01T12:00:00Z"
        type: string
      approved_by:
        description: |-
          The investigation Approved By.
          example: 1
        type: integer
      auto_approval_failure_reason:
        type: string
      barcodes:
        description: |-
          Barcodes of the investigation.
          example: "1234567890"
        type: string
      department:
        description: |-
          The investigation Department of the investigation result.
          example: "Hematology"
        type: string
      device_value:
        description: |-
          The investigation device value of the investigation result.
          example: "12.5"
        type: string
      entered_at:
        description: |-
          The investigation result Entered At.
          example: "2021-07-01T12:00:00Z"
        type: string
      entered_by:
        description: |-
          The investigation Entered Technician.
          example: 1
        type: integer
      id:
        description: |-
          The investigation result ID.
          example: 1
        type: integer
      investigation_data:
        allOf:
        - $ref: '#/definitions/structures.InvestigationData'
        description: "The investigation Investigation Data.\nexample: {\n\t\t\"id\":
          1,\n\t\t\"data\": \"This is a investigation value in text\",\n\t\t\"data_type\":
          \"investigation_value_text\",\n\t\t\"created_at\": \"2021-07-01T12:00:00Z\",\n\t\t\"updated_at\":
          \"2021-07-01T12:00:00Z\",\n\t\t\"deleted_at\": \"2021-07-01T12:00:00Z\",\n\t\t\"created_by\":
          1,\n\t\t\"updated_by\": 1,\n\t\t\"deleted_by\": 1\n\t}"
      investigation_name:
        description: |-
          The investigation name of the investigation result.
          example: "Hemoglobin"
        type: string
      investigation_value:
        description: |-
          The investigation value of the investigation result.
          example: "12.5"
        type: string
      is_auto_approved:
        description: |-
          If the investigation was auto approved.
          example: true
        type: boolean
      is_critical:
        description: |-
          The investigation Is Critical.
          example: true
        type: boolean
      is_non_reportable:
        description: |-
          The investigation Reportable in PDF.
          example: true
        type: boolean
      lis_code:
        description: |-
          The investigation Lis-Code.
          example: "HGB0001"
        type: string
      master_investigation_id:
        description: |-
          The investigation ID of the investigation result.
          example: 1
        type: integer
      master_test_id:
        description: |-
          The investigation Master Test ID.
          example: 1
        type: integer
      method:
        description: |-
          The investigation method performed.
          example: "Manual"
        type: string
      method_type:
        description: |-
          The investigation method type.
          example: "Manual"
        type: string
      reference_range_text:
        description: |-
          The investigation reference range of the investigation result.
          example: "12.0 - 16.0"
        type: string
      result_representation_type:
        description: |-
          The investigation result representation type of the investigation result.
          example: "biopattern"
        type: string
      source:
        description: |-
          The investigation Source.
          example: "OH Lab BLR"
        type: string
      status:
        description: |-
          The investigation Status.
          example: "Approved"
        type: string
      test_detail_id:
        description: |-
          The test detail ID of the investigation result.
          example: 1
        type: integer
      test_details_status:
        description: |-
          The test detail status of the investigation result.
          example: "approved"
        type: string
      uom:
        description: |-
          The investigation unit of measurement of the investigation result.
          example: "g/dL"
        type: string
    type: object
...
  structures.Log:
    properties:
      request_logs:
        description: The request details of the Log.
        items:
          $ref: '#/definitions/structures.RequestLog'
        type: array
    type: object
  structures.ModifyValueApiRequest:
    properties:
      current_investigation:
        $ref: '#/definitions/structures.ModifyValueRequest'
      past_investigations:
        items:
          $ref: '#/definitions/structures.ModifyValueRequest'
        type: array
    type: object
  structures.ModifyValueRequest:
    properties:
      investigation_value:
        type: string
      master_investigation_id:
        type: integer
    type: object
  structures.OrderLogs:
    properties:
      id:
        description: |-
          The ID of the OrderLogs.
          example: 1
        type: string
      sample_logs:
        description: |-
          The sample logs of the RequestLog.
          example: [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "test_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "investigation_logs": [{"investigation_name": "CBC", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]}]}]
        items:
          $ref: '#/definitions/structures.SampleLogs'
        type: array
      status_logs:
        description: |-
          The status logs of the OrderLogs.
          example: [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.StatusLog'
        type: array
    type: object
...
  structures.PatientPastRecordsApiResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structures.PatientPastRecordsData'
        type: array
      dates:
        items:
          type: string
        type: array
      patient_details:
        $ref: '#/definitions/structures.PatientPastRecordsPatientDetails'
    type: object
  structures.PatientPastRecordsData:
    properties:
      department:
        type: string
      investigations:
        items:
          $ref: '#/definitions/structures.PatientPastRecordsInvestigations'
        type: array
    type: object
  structures.PatientPastRecordsInvestigations:
    properties:
      investigation_id:
        type: integer
      investigation_name:
        type: string
      past_records:
        additionalProperties:
          items:
            $ref: '#/definitions/structures.PatientPastRecordsValue'
          type: array
        type: object
    type: object
  structures.PatientPastRecordsPatientDetails:
    properties:
      age_days:
        type: integer
      age_months:
        type: integer
      age_years:
        type: integer
      gender:
        type: string
      name:
        type: string
    type: object
  structures.PatientPastRecordsValue:
    properties:
      investigation_data:
        type: string
      investigation_value:
        type: string
      uom:
        type: string
    type: object
  structures.RequestLog:
    properties:
      id:
        description: |-
          The ID of the RequestLog.
          example: 1
        type: string
      order_logs:
        description: |-
          The order logs of the RequestLog.
          example: [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "sample_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "test_logs": [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "investigation_logs": [{"investigation_name": "CBC", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]}]}]}]
        items:
          $ref: '#/definitions/structures.OrderLogs'
        type: array
      status_logs:
        description: |-
          The status logs of the RequestLog.
          example: [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.StatusLog'
        type: array
    type: object
  structures.SampleLogs:
    properties:
      id:
        description: |-
          The ID of the SampleLogs.
          example: 1
        type: string
      status_logs:
        description: |-
          The status logs of the SampleLogs.
          example: [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.StatusLog'
        type: array
      test_logs:
        description: |-
          The test logs of the SampleLogs.
          example: [{"id": "1", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}], "investigation_logs": [{"investigation_name": "CBC", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]}]
        items:
          $ref: '#/definitions/structures.TestLogs'
        type: array
    type: object
  structures.StatusLog:
    properties:
      status:
        description: |-
          The status of the StatusLog.
          example: "Pending"
        type: string
      timestamp:
        description: |-
          The timestamp of the StatusLog.
          example: "2021-07-01T12:00:00Z"
        type: string
      user:
        description: |-
          The user of the StatusLog.
          example: "John Doe"
        type: string
      user_id:
        description: |-
          The user ID of the StatusLog.
          example: "1"
        type: string
    type: object
...
  structures.TestLogs:
    properties:
      id:
        description: |-
          The ID of the TestLogs.
          example: 1
        type: string
      investigation_logs:
        description: |-
          The investigation logs of the TestLogs.
          example: [{"investigation_name": "CBC", "status_logs": [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.InvestigationLogs'
        type: array
      status_logs:
        description: |-
          The status logs of the TestLogs.
          example: [{"status": "Pending", "timestamp": "2021-07-01T12:00
          :00Z", "user": "John Doe", "user_id": "1"}]
        items:
          $ref: '#/definitions/structures.StatusLog'
        type: array
    type: object
...
info:
  contact: {}
  description: Citadel API Documentation
  title: Citadel API
paths:
...
  /api/v1/audit_logs/tasks/{taskId}:
    get:
      description: Get all audit logs by order id
      parameters:
      - description: Task ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit Logs
          schema:
            items:
              $ref: '#/definitions/structures.AuditLog'
            type: array
        "400":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
      summary: Get Audit Logs
      tags:
      - audit_logs
  /api/v1/investigation-results/patients/delta-values:
    get:
      description: Get delta values from patient id
      parameters:
      - description: Patient Id
        in: query
        name: patient_id
        required: true
        type: integer
      - description: Master Investigation Ids
        in: query
        name: master_investigation_ids
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delta Values
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/structures.DeltaValuesStructResponse'
              type: array
            type: object
      summary: Get Delta Values from Patient ID
      tags:
      - investigation-results
  /api/v1/investigation-results/patients/past-records:
    get:
      description: Get all investigation results of a patient
      parameters:
      - description: Patient ID
        in: path
        name: patient_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Patient Past Records
          schema:
            $ref: '#/definitions/structures.PatientPastRecordsApiResponse'
      summary: Get Patient Past Records
      tags:
      - investigation-results
  /api/v1/investigation-results/tasks/{taskId}/details:
    get:
      description: Get all investigation results by task id
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Investigation Results
          schema:
            items:
              $ref: '#/definitions/structures.InvestigationResult'
            type: array
      summary: Get Investigation Results
      tags:
      - investigation-results
  /api/v1/investigation-results/tasks/{taskId}/investigation-abnormality:
    get:
      description: Get investigation abnormality by task id, investigation code and
        value
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Investigation Code
        in: query
        name: investigation_code
        required: true
        type: string
      - description: Investigation Value
        in: query
        name: investigation_value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Abnormality
          schema:
            $ref: '#/definitions/structures.CommonAPIResponse'
      summary: Get Investigation Abnormality
      tags:
      - investigation-results
  /api/v1/investigation-results/tasks/{taskId}/modify-value:
    post:
      description: Get derived investigations and abnormality by task id, investigation
        code and value
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Modify Value API Request
        in: body
        name: ModifyValueApiRequest
        required: true
        schema:
          $ref: '#/definitions/structures.ModifyValueApiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Abnormality
          schema:
            $ref: '#/definitions/structures.CommonAPIResponse'
      summary: Get Derived Investigations and Abnormality
      tags:
      - investigation-results
...
  /api/v1/task-metadata/task/{taskId}/details:
    get:
      description: Get Task Metadata Details for the given Task ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Common API Response
          schema:
            $ref: '#/definitions/github_com_Orange-Health_citadel_apps_task_metadata_structures.TaskMetadata'
        "400":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "404":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "500":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
      summary: GetTaskMetadataDetails
      tags:
      - task-metadata
  /api/v1/task-metadata/task/{taskId}/last-event-sent-at:
    patch:
      description: Update Last Event Sent At for the given Task ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Common API Response
          schema:
            type: string
        "400":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "404":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "500":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
      summary: UpdateLastEventSentAt
      tags:
      - task-metadata
  /api/v1/task-metadata/tasks/:taskId:
    patch:
      consumes:
      - application/json
      description: Update Task Metadata Details for the given Task ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Task Metadata Object
        in: body
        name: taskMetadata
        required: true
        schema:
          $ref: '#/definitions/models.TaskMetadata'
      produces:
      - application/json
      responses:
        "200":
          description: Common API Response
          schema:
            $ref: '#/definitions/models.TaskMetadata'
        "400":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "404":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
        "500":
          description: Common API Response
          schema:
            $ref: '#/definitions/structures.CommonError'
      summary: UpdateTaskMetadata
      tags:
      - task-metadata
...
swagger: "2.0"